Only split out ec and dsa
RSA is being mvoed out of the interfaces namespace in PR #1592.
